# perf: improve proving performance