--- a/src/groth16/mapped_params.rs
+++ b/src/groth16/mapped_params.rs
@@ -6,7 +6,6 @@ use rayon::prelude::*;
 use crate::SynthesisError;
 
 use memmap::Mmap;
-use rayon::prelude::*;
 
 use std::fs::File;
 use std::io;