--- a/src/groth16/prover.rs
+++ b/src/groth16/prover.rs
@@ -287,6 +287,9 @@ where
     let input_len = input_assignments[0].len();
     let vk = params.get_vk(input_len)?.clone();
     let n = provers[0].a.len();
+    let a_aux_density_total = provers[0].a_aux_density.get_total_density();
+    let b_input_density_total = provers[0].b_input_density.get_total_density();
+    let b_aux_density_total = provers[0].b_aux_density.get_total_density();
 
     // Make sure all circuits have the same input len.
     for prover in &provers {
@@ -294,6 +297,21 @@ where
             prover.a.len(),
             n,
             "only equaly sized circuits are supported"
+        );
+        assert_eq!(
+            a_aux_density_total,
+            prover.a_aux_density.get_total_density(),
+            "only identical circuits are supported"
+        );
+        assert_eq!(
+            b_input_density_total,
+            prover.b_input_density.get_total_density(),
+            "only identical circuits are supported"
+        );
+        assert_eq!(
+            b_aux_density_total,
+            prover.b_aux_density.get_total_density(),
+            "only identical circuits are supported"
         );
     }
 
@@ -363,12 +381,15 @@ where
     drop(fft_kern);
     let mut multiexp_kern = Some(LockedMultiexpKernel::<E>::new(log_d, priority));
 
+    let params_h = params.get_h(n)?;
+    let params_l = params.get_l(provers.len())?;
+
     let h_s = a_s
         .into_iter()
         .map(|a| {
             let h = multiexp(
                 &worker,
-                params.get_h(a.len())?,
+                params_h.clone(),
                 FullDensity,
                 a,
                 &mut multiexp_kern,
@@ -382,7 +403,7 @@ where
         .map(|aux_assignment| {
             let l = multiexp(
                 &worker,
-                params.get_l(aux_assignment.len())?,
+                params_l.clone(),
                 FullDensity,
                 aux_assignment.clone(),
                 &mut multiexp_kern,
@@ -391,19 +412,20 @@ where
         })
         .collect::<Result<Vec<_>, SynthesisError>>()?;
 
+    let (a_inputs_source, a_aux_source) = params.get_a(input_len, a_aux_density_total)?;
+    let (b_g1_inputs_source, b_g1_aux_source) =
+        params.get_b_g1(b_input_density_total, b_aux_density_total)?;
+    let (b_g2_inputs_source, b_g2_aux_source) =
+        params.get_b_g2(b_input_density_total, b_aux_density_total)?;
+
     let inputs = provers
         .into_iter()
         .zip(input_assignments.iter())
         .zip(aux_assignments.iter())
         .map(|((prover, input_assignment), aux_assignment)| {
-            let a_aux_density_total = prover.a_aux_density.get_total_density();
-
-            let (a_inputs_source, a_aux_source) =
-                params.get_a(input_assignment.len(), a_aux_density_total)?;
-
             let a_inputs = multiexp(
                 &worker,
-                a_inputs_source,
+                a_inputs_source.clone(),
                 FullDensity,
                 input_assignment.clone(),
                 &mut multiexp_kern,
@@ -411,23 +433,18 @@ where
 
             let a_aux = multiexp(
                 &worker,
-                a_aux_source,
+                a_aux_source.clone(),
                 Arc::new(prover.a_aux_density),
                 aux_assignment.clone(),
                 &mut multiexp_kern,
             );
 
             let b_input_density = Arc::new(prover.b_input_density);
-            let b_input_density_total = b_input_density.get_total_density();
             let b_aux_density = Arc::new(prover.b_aux_density);
-            let b_aux_density_total = b_aux_density.get_total_density();
-
-            let (b_g1_inputs_source, b_g1_aux_source) =
-                params.get_b_g1(b_input_density_total, b_aux_density_total)?;
 
             let b_g1_inputs = multiexp(
                 &worker,
-                b_g1_inputs_source,
+                b_g1_inputs_source.clone(),
                 b_input_density.clone(),
                 input_assignment.clone(),
                 &mut multiexp_kern,
@@ -435,25 +452,22 @@ where
 
             let b_g1_aux = multiexp(
                 &worker,
-                b_g1_aux_source,
+                b_g1_aux_source.clone(),
                 b_aux_density.clone(),
                 aux_assignment.clone(),
                 &mut multiexp_kern,
             );
 
-            let (b_g2_inputs_source, b_g2_aux_source) =
-                params.get_b_g2(b_input_density_total, b_aux_density_total)?;
-
             let b_g2_inputs = multiexp(
                 &worker,
-                b_g2_inputs_source,
+                b_g2_inputs_source.clone(),
                 b_input_density,
                 input_assignment.clone(),
                 &mut multiexp_kern,
             );
             let b_g2_aux = multiexp(
                 &worker,
-                b_g2_aux_source,
+                b_g2_aux_source.clone(),
                 b_aux_density,
                 aux_assignment.clone(),
                 &mut multiexp_kern,