--- a/src/groth16/prover.rs
+++ b/src/groth16/prover.rs
@@ -322,25 +322,40 @@ where
             let mut c =
                 EvaluationDomain::from_coeffs(std::mem::replace(&mut prover.c, Vec::new()))?;
 
-            a.ifft(&worker, &mut fft_kern)?;
-            a.coset_fft(&worker, &mut fft_kern)?;
-            b.ifft(&worker, &mut fft_kern)?;
-            b.coset_fft(&worker, &mut fft_kern)?;
-            c.ifft(&worker, &mut fft_kern)?;
-            c.coset_fft(&worker, &mut fft_kern)?;
+            // execute 1 fft on the CPU, and 2 on the GPU if enabled
+            let (a_res, bc_res) = rayon::join(
+                || {
+                    a.ifft(&worker, &mut None)?;
+                    a.coset_fft(&worker, &mut None)
+                },
+                || {
+                    b.ifft(&worker, &mut fft_kern)?;
+                    b.coset_fft(&worker, &mut fft_kern)?;
+
+                    c.ifft(&worker, &mut fft_kern)?;
+                    c.coset_fft(&worker, &mut fft_kern)
+                },
+            );
+            a_res?;
+            bc_res?;
 
             a.mul_assign(&worker, &b);
-            drop(b);
             a.sub_assign(&worker, &c);
+
+            drop(b);
             drop(c);
+
             a.divide_by_z_on_coset(&worker);
             a.icoset_fft(&worker, &mut fft_kern)?;
-            let mut a = a.into_coeffs();
+
+            let a = a.into_coeffs();
             let a_len = a.len() - 1;
-            a.truncate(a_len);
 
             Ok(Arc::new(
-                a.into_iter().map(|s| s.0.into_repr()).collect::<Vec<_>>(),
+                a.into_iter()
+                    .take(a_len)
+                    .map(|s| s.0.into_repr())
+                    .collect::<Vec<_>>(),
             ))
         })
         .collect::<Result<Vec<_>, SynthesisError>>()?;