- parallelize loading mapped params
- load params in parallel while GPU work is being done, to avoid blocking later
- improve parallelism inside proving code to avoid bottlenecks

## Speed Improvements

on a benchmark machine using 1 RTX 3090 and 1 RTX 2080TI

- Generate PoRep: before: `2,283s`, after: `881s`
- Generate Window PoST: before: `321s`, after: `144s`