@dignifiedquire 2021-02-17
I expanded the thread pool install but now I am getting

```
│2021-02-17T15:11:29.056 ERROR bellperson::multicore > The wait call should never be done inside the worker thread pool
```