@dignifiedquire 2021-02-17
It is not inside the install call, but inside a `par_iter`. Will test if this has perf impact if I remove it