@cryptonemo 2021-02-17
> > It is not inside the install call, but inside a `par_iter`. Will test if this has perf impact if I remove it
> 
> If the wait is inside the par_iter, which is inside of the install call, then yes, it's in the install ... and can't happen like that :-)

Disabling the par_iter usage isn't the issue, as a wait is still called inside of the `install` context.  One way to resolve this is to re-factor the code such that the waiter(s) is(are) returned and the waits are called after the `install` execution is complete.