@cryptonemo 2021-02-17
Generally looks better.  Is it definitely the case that the usage of `par!` (the main caller of `spawn` now) is constrained to something reasonable like max cpus?  If not, we still run the risk of crashing due to pending spawn requests.