@dignifiedquire 2021-02-18
`par` uses the default rayon threadpool out of the box, or the one that is currently `install`ed. The default threadpool uses number of cpus as the default constraint