@cryptonemo 2021-02-18
I tried running this code and it hangs, which is consistent with the CI timeouts.  If you don't see the hang, try lowering the `BELLMAN_NUM_CPU` value.