@cryptonemo 2021-02-18
Just throwing this out there to think about.  There's a ton of noise and re-factoring in this PR and it's obviously causing logic errors and adding complexity.  The main goal isn't clear and is difficult to review in this PR.  The secondary goal is trivially implemented as this diff on top of master.  I don't want to add work for anyone, but would appreciate an alternative PR that more minimally illustrates the FFT CPU/GPU split if possible.

```
diff --git a/src/groth16/mapped_params.rs b/src/groth16/mapped_params.rs
index 4d39b9a..3c6d722 100644
--- a/src/groth16/mapped_params.rs
+++ b/src/groth16/mapped_params.rs
@@ -4,6 +4,7 @@ use groupy::{CurveAffine, EncodedPoint};
 use crate::SynthesisError;
 
 use memmap::Mmap;
+use rayon::prelude::*;
 
 use std::fs::File;
 use std::io;
@@ -59,7 +60,7 @@ impl<'a, E: Engine> ParameterSource<E> for &'a MappedParameters<E> {
     fn get_h(&self, _num_h: usize) -> Result<Self::G1Builder, SynthesisError> {
         let builder = self
             .h
-            .iter()
+            .par_iter()
             .cloned()
             .map(|h| read_g1::<E>(&self.params, h, self.checked))
             .collect::<Result<_, _>>()?;
@@ -70,7 +71,7 @@ impl<'a, E: Engine> ParameterSource<E> for &'a MappedParameters<E> {
     fn get_l(&self, _num_l: usize) -> Result<Self::G1Builder, SynthesisError> {
         let builder = self
             .l
-            .iter()
+            .par_iter()
             .cloned()
             .map(|l| read_g1::<E>(&self.params, l, self.checked))
             .collect::<Result<_, _>>()?;
@@ -85,7 +86,7 @@ impl<'a, E: Engine> ParameterSource<E> for &'a MappedParameters<E> {
     ) -> Result<(Self::G1Builder, Self::G1Builder), SynthesisError> {
         let builder = self
             .a
-            .iter()
+            .par_iter()
             .cloned()
             .map(|a| read_g1::<E>(&self.params, a, self.checked))
             .collect::<Result<_, _>>()?;
@@ -102,7 +103,7 @@ impl<'a, E: Engine> ParameterSource<E> for &'a MappedParameters<E> {
     ) -> Result<(Self::G1Builder, Self::G1Builder), SynthesisError> {
         let builder = self
             .b_g1
-            .iter()
+            .par_iter()
             .cloned()
             .map(|b_g1| read_g1::<E>(&self.params, b_g1, self.checked))
             .collect::<Result<_, _>>()?;
@@ -119,7 +120,7 @@ impl<'a, E: Engine> ParameterSource<E> for &'a MappedParameters<E> {
     ) -> Result<(Self::G2Builder, Self::G2Builder), SynthesisError> {
         let builder = self
             .b_g2
-            .iter()
+            .par_iter()
             .cloned()
             .map(|b_g2| read_g2::<E>(&self.params, b_g2, self.checked))
             .collect::<Result<_, _>>()?;

```