@dignifiedquire 2021-02-18
@cryptonemo unfortunately the original description does not cover the actual changes anymore, the different refactors inside `prover.rs` are all needed to actually get the new performance to happen.