@cryptonemo 2021-02-18
As a side note, I think we could eventually remove the Worker/Waiter entirely, but that should be a separate PR.

> @cryptonemo unfortunately the original description does not cover the actual changes anymore, the different refactors inside `prover.rs` are all needed to actually get the new performance to happen.

Put another way, I suspect I could do this more minimally, but don't want to re-create work unless it's a high priority to resolve these issues.