@dignifiedquire 2021-02-18
> The `pub struct Waiter` can completely be removed.

done