@dignifiedquire 2021-02-18
> I suspect I could do this more minimally

Could you elaborate? Both the usage of the `par` macro (or the equivalent inlined spawn calls) and the removal of the `Waiter` construction are required for perf, so I am not sure what you think can be reduced.