@cryptonemo 2021-02-18
> > I suspect I could do this more minimally
> 
> Could you elaborate? Both the usage of the `par` macro (or the equivalent inlined spawn calls) and the removal of the `Waiter` construction are required for perf, so I am not sure what you think can be reduced.

As an example, I think the `par` macro makes the parallelism less readable, or rather I find it confusing to know what it's doing in the runtime context that it's in (i.e. is it inside of an install or not?  Should it be?)  I'm not saying that inlining a spawn or using a pool compute is more readable, but I also have a handle on what we have on master and find these changes confusing.  I'm reading diffs on github though, so the code may be a little more clear -- but given that it adds hangs, I suspect there's more subtle errors still present (until resolved).