@cryptonemo 2021-02-18
> > but given that it adds hangs, I suspect there's more subtle errors still present (until resolved).
> 
> There are definitely issues with the tests, though it works great when running tests individually and when running it through `benchy`.

Maybe try running tests locally with `BELLMAN_NUM_CPUS=1`?  That was a very helpful way to debug previous hangs.