@dignifiedquire 2021-07-23
rebased onto #204